ajuste no menu

+++ trunk/fonte/ieadam/src/main/java/br/com/ieadam/repositorio/UsuarioPerfilRepositorio.java
@@ -7,6 +7,6 @@ import br.com.ieadam.dominio.UsuarioPerfil;
 
 @Repository
 public interface UsuarioPerfilRepositorio extends
-		JpaRepository<UsuarioPerfil, Integer> {
+		JpaRepository<UsuarioPerfil, Long> {
 
 }
